Recompile tier1.lib for 🪟 Fix size_t warnings Update LZMA to 2019-02-02 - 2013 LZMA is uncompilable on 64-bit and honestly I CBA to figure out the exact version used by the beta, so just use the latest one. This 17 year old game is getting Vulcan rendering for goodness sake. Fix X64BITS being used in CUtlBuffer::GetPtr  - Supposed to be PLATFORM_64BITS Fix warnings about long being truncated to int Use V_strlen over regular strlen and other functions using int's

+++ tier1/KeyValues.cpp
@@ -1872,7 +1872,7 @@ void KeyValues::CopyKeyValue( const KeyValues& src, size_t tmpBufferSizeB, char*
 	case TYPE_INT:
 		{
 			m_iValue = src.m_iValue;
-			Q_snprintf( tmpBuffer, tmpBufferSizeB, "%d", m_iValue );
+			Q_snprintf( tmpBuffer, (int)tmpBufferSizeB, "%d", m_iValue );
 			int len = Q_strlen(tmpBuffer) + 1;
 				m_sValue = KVStringAlloc<char>(len);
 			Q_strncpy( m_sValue, tmpBuffer, len  );
@@ -1881,7 +1881,7 @@ void KeyValues::CopyKeyValue( const KeyValues& src, size_t tmpBufferSizeB, char*
 	case TYPE_FLOAT:
 		{
 			m_flValue = src.m_flValue;
-			Q_snprintf( tmpBuffer, tmpBufferSizeB, "%f", m_flValue );
+			Q_snprintf( tmpBuffer, (int)tmpBufferSizeB, "%f", m_flValue );
 			int len = Q_strlen(tmpBuffer) + 1;
 			m_sValue = KVStringAlloc<char>(len);
 			Q_strncpy( m_sValue, tmpBuffer, len );
